openpilot v0.9.4 release
date: 2023-07-27T18:38:32
master commit: fa310d9e2542cf497d92f007baec8fd751ffa99c

+++ panda/board/safety_declarations.h
@@ -67,7 +67,7 @@ typedef struct {
   const bool has_steer_req_tolerance;
 
   // angle cmd limits
-  const int angle_deg_to_can;
+  const float angle_deg_to_can;
   const struct lookup_t angle_rate_up_lookup;
   const struct lookup_t angle_rate_down_lookup;
   const int max_angle_error;             // used to limit error between meas and cmd while enabled
